packages/server-python/requirements.txt
attrs==20.2.0
autopep8==1.5.4
click==7.1.2
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytest-asyncio==0.14.0
python-dotenv==0.14.0
rsa==4.6
six==1.15.0
Telethon==1.16.4
toml==0.10.1
Werkzeug==1.0.1
